refactor(codebuild-build-batch): convert to libnuke resource format
